Rename the file causing GithubActions to fail
