+++ app/src/components/blocks/TextBlockComponent.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import {BlockDataStore} from 'src/db/BlockDataStore'
+import {BlockDataStore} from 'src/database/BlockDataStore'
 import {useDb} from 'src/providers/DbProvider'
 import {TextBlock} from 'src/types/blocks/TextBlock'
 import {Contenteditable} from '../contenteditable/Contenteditable'
