+++ app/src/database/BlockDataStore.ts
@@ -1,10 +1,10 @@
 import {Block} from '../types/Block'
 import {TableName} from '../types/TablesName'
-import {Db} from './Db'
+import {DatabaseManager} from './DatabaseManager'
 import {BlockType} from 'src/types/BlockType'
 
 export class BlockDataStore {
-  static create(db: Db): void {
+  static create(db: DatabaseManager): void {
     const objectStore = db
       .instance()
       .createObjectStore(TableName.Block, {keyPath: 'id'})
@@ -14,19 +14,32 @@ export class BlockDataStore {
     objectStore.createIndex('type', 'type', {unique: false})
   }
 
-  static async insert<T extends Block>(db: Db, block: T): Promise<void> {
+  static async insert<T extends Block>(
+    db: DatabaseManager,
+    block: T,
+  ): Promise<void> {
     await db.insert<T>(TableName.Block, block)
   }
 
-  static getById<T extends Block>(db: Db, id: string): Promise<T | undefined> {
+  static getById<T extends Block>(
+    db: DatabaseManager,
+    id: string,
+  ): Promise<T | undefined> {
     return db.getById(TableName.Block, id)
   }
 
-  static getAllByType<T extends Block>(db: Db, type: BlockType): Promise<T[]> {
+  static getAllByType<T extends Block>(
+    db: DatabaseManager,
+    type: BlockType,
+  ): Promise<T[]> {
     return db.getAllByIndex<T>(TableName.Block, 'type', type)
   }
 
-  static updateById(db: Db, id: string, block: Block): Promise<void> {
+  static updateById(
+    db: DatabaseManager,
+    id: string,
+    block: Block,
+  ): Promise<void> {
     return db.updateById<Block>(TableName.Block, id, block)
   }
 }
