+++ app/src/screens/HomeScreen.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import {Link} from 'react-router-dom'
-import {usePages} from 'src/db/queryHooks'
+import {usePages} from 'src/database/queryHooks'
 import {WelcomeScreen} from './WelcomeScreen'
 
 export const HomeScreen: React.FC = () => {
